code/requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
astroid==2.1.0
asymmetric-jwt-auth==0.4.3
Babel==2.6.0
certbot-django==0.2.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
defusedxml==0.5.0
diff-match-patch==20181111
dj-static==0.0.6
Django==2.1.4
django-bootstrap-modal-forms==1.3.1
django-bootstrap4==0.0.7
django-debug-toolbar==1.11
django-import-export==1.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
docutils==0.14
et-xmlfile==1.0.1
idna==2.8
imagesize==1.1.0
isort==4.3.4
jdcal==1.4
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
odfpy==1.4.0
openpyxl==2.5.12
packaging==19.0
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2-binary==2.7.6.1
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.2.2
pyparsing==2.4.0
python-decouple==3.1
pytz==2018.7
PyYAML==3.13
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.2.4
static3==0.7.0
tablib==0.12.1
typed-ast==1.1.1
unicodecsv==0.14.1
urllib3==1.24.2
wrapt==1.10.11
xlrd==1.2.0
xlwt==1.3.0
